requirements.txt
python-telegram-bot==20.7
pillow==10.2.0
pytesseract==0.3.10
openai==1.3.0
python-dotenv==1.0.0